pyproject.toml
[tool.poetry]
name = "moodle-1"
version = "0.1.0"
description = ""
authors = ["repl.it user <replituser@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.7"
django-countries = "^7.0"
django-taggit = "^1.3.0"
django-crispy-forms = "^1.11"
django-tinymce = "^3.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
